+++ src/main/clojure/clojure/core/matrix/impl/dataset.clj
@@ -10,28 +10,36 @@
             [clojure.core.matrix.impl defaults persistent-vector] ;; these are needed during loading
             [clojure.core.matrix.protocols :as mp]
             [clojure.core.matrix.macros :refer [error]]
-            [clojure.core.matrix.utils])
+            [clojure.core.matrix.utils :as utils])
   (:import [java.io Writer])
-  (:import [clojure.lang IPersistentVector]
+  (:import [clojure.lang IPersistentVector IPersistentMap]
            [java.util List]))
 
 (set! *warn-on-reflection* true)
 (set! *unchecked-math* :warn-on-boxed)
 
-;; a column-based DataSet implementation.
-;; columns are arbitrary core.matrix arrays, treated as vectors
+(declare wrap-row)
+
+;; A class to represent a single dataset row, as a reference into a vector of columns
 
 (deftype DataSetRow
   [^IPersistentVector column-names
    ^IPersistentVector columns
    ^long index]
+  clojure.lang.IMeta
+    (meta [m]
+      nil)
+  clojure.lang.IObj
+    (withMeta [m meta]
+      (with-meta (mp/convert-to-nested-vectors m) meta))
   clojure.lang.Indexed
     (count [m] (count column-names))
     (nth [m i]
       (mp/get-1d (nth columns i) index))
     (nth [m i not-found]
-      (when (< -1 i (long (mp/dimension-count (first columns) 0)))
-        (mp/get-1d (nth columns i) index)))
+      (if (< -1 i (long (mp/dimension-count (first columns) 0)))
+        (mp/get-1d (nth columns i) index)
+        not-found))
   clojure.lang.Seqable
     (seq [m]
       (seq (mp/convert-to-nested-vectors m)))
@@ -47,17 +55,64 @@
     (assocN [m i v]
       (assoc (mp/convert-to-nested-vectors m) i v)))
 
-(defn- wrap-row [dataset index]
-  (let [index (long index)
-        cols (mp/get-columns dataset)]
-    (when-not (< -1 index (long (mp/dimension-count (first cols) 0)))
-      (error "Row index does not exist: " index))
-    (DataSetRow. (mp/column-names dataset) cols index)))
+(defn wrap-row
+  "Wraps a row of a datset as a DataSetRow"
+  ([dataset index]
+    (let [index (long index)
+          cols (mp/get-columns dataset)]
+      (when-not (< -1 index (long (mp/dimension-count (first cols) 0)))
+        (error "Row index does not exist: " index))
+      (DataSetRow. (mp/column-names dataset) cols index)))
+  ([dataset index not-found]
+    (let [index (long index)
+          cols (mp/get-columns dataset)]
+      (if (< -1 index (long (mp/dimension-count (first cols) 0)))
+        (DataSetRow. (mp/column-names dataset) cols index)
+        not-found))))
 
-(defrecord DataSet
+;; a column-based DataSet implementation.
+;; columns are arbitrary core.matrix arrays, treated as vectors
+
+(deftype DataSet
   [^IPersistentVector column-names
    ^IPersistentVector columns
-   ^IPersistentVector shape])
+   ^IPersistentVector shape]
+  clojure.lang.IMeta
+    (meta [m]
+      nil)
+  clojure.lang.IObj
+    (withMeta [m meta]
+      (with-meta (mp/convert-to-nested-vectors m) meta))
+  clojure.lang.Indexed
+    (count [m] (first shape))
+    (nth [m i]
+      (wrap-row m i))
+    (nth [m i not-found]
+      (wrap-row m i not-found))
+  clojure.lang.Seqable
+    (seq [m]
+      (map #(wrap-row m %) (range (first shape))))
+  clojure.lang.ILookup
+    (valAt [m i]
+      (wrap-row m i))
+    (valAt [m i not-found]
+      (wrap-row m i not-found))
+  clojure.lang.IFn
+    (invoke [m i]
+      (wrap-row m i))
+    (invoke [m i not-found]
+      (wrap-row m i not-found))
+  clojure.lang.IPersistentCollection
+    (cons [m v]
+      (conj (mp/convert-to-nested-vectors m) v))
+    (empty [m]
+      [])
+    (equiv [m a]
+      (= (mp/convert-to-nested-vectors m) a))
+  clojure.lang.IPersistentVector
+    (length [m] (first shape))
+    (assocN [m i v]
+      (assoc (mp/convert-to-nested-vectors m) i v)))
 
 (defn dataset-from-columns [col-names cols]
   (let [^IPersistentVector col-names (vec col-names)
@@ -81,7 +136,9 @@
     (DataSet. col-names cols [rc cc])))
 
 (defn dataset-from-array
-  "Construct a dataset from an array. Uses labels from the array if available."
+  "Construct a dataset from an array. 
+
+   Uses labels from the array as column names if available, or long values 0, 1.... if not."
   ([m]
     (let [dims (long (mp/dimensionality m))]
       (when (not= dims 2)
@@ -103,13 +160,13 @@
       (let [dim (long dim)
             i (long i)]
         (cond
-          (== dim 1) (nth (:column-names m) i)
+          (== dim 1) (nth (.column-names m) i)
           (<= 0 (long i) (dec (long (mp/dimension-count m dim)))) nil
           :else (error "Dimension index out of range: " i))))
     (labels [m dim]
       (let [dim (long dim)]
         (cond
-          (== dim 1) (:column-names m)
+          (== dim 1) (.column-names m)
           (<= 0 (long dim) (dec (long (mp/dimensionality m)))) nil
           :else (error "Dimension out of range: " dim)))))
 
@@ -127,12 +184,29 @@
   (get-major-slice [ds i]
     (mp/get-major-slice (mp/get-rows ds) i))
   (get-slice [ds dimension i]
-    (mp/get-slice (mp/get-rows ds) dimension i)))
+    (mp/get-slice (mp/get-rows ds) dimension i))
+  
+  DataSetRow
+  (get-column [ds i]
+    (mp/get-1d (nth (.columns ds) i) (.index ds)))
+  (get-row [ds i]
+    (error "Cannot get rows from a single DataSetRow"))
+  (get-major-slice [ds i]
+    (mp/get-1d (nth (.columns ds) i) (.index ds)))
+  (get-slice [ds dimension i]
+    (if (== (long dimension) 0)
+      (mp/get-1d (nth (.columns ds) i) (.index ds))
+      (error "Cannot get dimension " dimension " from DataSetRow"))))
 
 (extend-protocol mp/PMatrixColumns
   DataSet
     (get-columns [ds]
-      (.columns ds)))
+      (.columns ds))
+  DataSetRow
+    (get-columns [ds]
+      (let [cols (.columns ds)
+            ix (.index ds)]
+        (mapv #(mp/get-1d % ix) cols))))
 
 (extend-protocol mp/PMatrixRows
   DataSet
@@ -141,6 +215,24 @@
         (mapv
           (fn [i] (wrap-row ds i))
           (range row-count)))))
+
+(extend-protocol mp/PConversion
+  DataSet
+    (convert-to-nested-vectors [ds]
+      (let [cols (mapv mp/convert-to-nested-vectors (.columns ds))]
+        (mp/transpose cols)))
+  DataSetRow
+    (convert-to-nested-vectors [ds]
+      (let [cols (.columns ds)
+            ix (.index ds)]
+        (mapv #(mp/get-1d % ix) cols))))
+
+(extend-protocol mp/PColumnIndex
+  Object
+    (column-index [ds column-name]
+      (when-let [cnames (mp/column-names ds)]
+        (let [cnames ^IPersistentVector (vec cnames)]
+          (and cnames (utils/find-index cnames column-name))))))
 
 (extend-protocol mp/PColumnSetting
   DataSet
@@ -178,14 +270,6 @@
   (add-column [ds col-name col]
     (dataset-from-columns (conj (mp/column-names ds) col-name)
                           (conj (mp/get-columns ds) col)))
-  (row-maps [ds]
-    (let [col-names (mp/column-names ds)]
-      (map #(zipmap col-names %)
-           (mp/get-rows ds))))
-  (to-map [ds]
-    (into {} (map (fn [k v] [k v])
-                  (mp/column-names ds)
-                  (mp/get-columns ds))))
   (merge-datasets [ds1 ds2]
     (reduce
      (fn [acc [k v]]
@@ -231,6 +315,30 @@
          (concat (mp/column-names ds1) (mp/column-names ds2))
          (concat (mp/get-columns ds1) (mp/get-columns ds2)))
         (error "Duplicate column names: " intersection)))))
+
+(extend-protocol mp/PDatasetMaps
+  DataSet
+    (row-maps [ds]
+      (let [col-names (mp/column-names ds)]
+        (mapv #(zipmap col-names %)
+              (mp/get-rows ds))))
+    (to-map [ds]
+      (into {} (map (fn [k v] [k v])
+                    (mp/column-names ds)
+                    (mp/get-columns ds))))
+  
+  Object
+    (row-maps [ds]
+      (if-let [col-names (mp/column-names ds)]
+        (mapv #(zipmap col-names (mp/element-seq %))
+              (mp/get-rows ds))
+        (error "No column names available")))
+    (to-map [ds]
+      (if-let [col-names (mp/column-names ds)]
+        (into {} (map (fn [k v] [k v])
+                      (mp/column-names ds)
+                      (mp/get-columns ds)))
+        (error "No column names available"))))
 
 (extend-protocol mp/PDimensionLabels
   DataSet
@@ -366,12 +474,17 @@
 
 (defn- broadcast-cols
   "Broadcasts a value to a vector of column values, in the shape specified"
-  ^IPersistentVector [a [rows cols]]
-  (case (long (mp/dimensionality a))
-    0 (vec (repeat cols a))
-    1 (mp/convert-to-nested-vectors a)
-    2 (vec (mp/get-columns a))
-    (error "Can't broadcast argument of shape " (mp/get-shape a))))
+  ^IPersistentVector [a [rows cols :as shape]]
+  (let []
+    (case (long (mp/dimensionality a))
+      0 (vec (repeat cols (vec (repeat rows a))))
+      1 (if (== (long cols) (long (mp/dimension-count a 0)))
+          (vec (mapv #(repeat rows %) (mp/element-seq a)))
+          (error "Can't broadcast to shape: " shape))
+      2 (if (== (long rows) (long (mp/dimension-count a 0)))
+          (vec (mp/get-columns a))
+          (error "Can't broadcast to shape: " shape))
+      (error "Can't broadcast argument of shape: " (mp/get-shape a)))))
 
 (extend-protocol mp/PFunctionalOperations
   DataSet
@@ -379,27 +492,21 @@
       (apply concat (mp/get-rows m)))
     (element-map
       ([m f]
-        (reduce
-          (fn [m i]
-            (update-in m [:columns i] mp/element-map f))
-          m
-          (range (.nth ^IPersistentVector (.shape m) 1))))
+        (let [cols (mapv #(mp/element-map % f) (.columns m))]
+          (DataSet. (.column-names m) cols (.shape m))))
       ([m f a]
-        (let [as (broadcast-cols a (mp/get-shape m))]
-          (reduce
-            (fn [m i]
-              (update-in m [:columns i] mp/element-map f (nth as i)))
-            m
-            (range (mp/dimension-count m 1)))))
+        (let [as (broadcast-cols a (mp/get-shape m))
+              cols (mapv #(mp/element-map %1 f %2) (.columns m) as)]
+          (DataSet. (.column-names m) cols (.shape m))))
       ([m f a more]
         (let [shape (.shape m)
               as (broadcast-cols a shape)
-              mores (apply mapv vector (map #(broadcast-cols % shape) more))]
-          (reduce
-            (fn [m i]
-              (update-in m [:columns i] mp/element-map f (nth as i) (nth mores i)))
-            m
-            (range (second shape))))))
+              mores (apply mapv vector (map #(broadcast-cols % shape) more))
+              cols (mapv #(mp/element-map %1 f %2 %3) 
+                         (.columns m) 
+                         as
+                         mores)]
+          (DataSet. (.column-names m) cols (.shape m)))))
     (element-map!
       ([m f]
         (mp/assign! m (mp/element-map m f)))
@@ -411,7 +518,34 @@
       ([m f]
         (reduce f (mp/element-seq m)))
       ([m f init]
-        (reduce f init (mp/element-seq m)))))
+        (reduce f init (mp/element-seq m))))
+    
+  DataSetRow
+    (element-seq [m]
+      m)
+    (element-map
+      ([m f]
+        (mapv f m))
+      ([m f a]
+        (let [as (mp/broadcast a (mp/get-shape m))]
+          (mapv f m (mp/element-seq a))))
+      ([m f a more]
+        (let [shape (mp/get-shape m)
+              as (mp/broadcast a shape)
+              mores (mapv #(mp/element-seq (mp/broadcast % shape)) more)]
+          (apply mapv f m (mp/element-seq a) mores))))
+    (element-map!
+      ([m f]
+        (mp/assign! m (mp/element-map m f)))
+      ([m f a]
+        (mp/assign! m (mp/element-map m f a)))
+      ([m f a more]
+        (mp/assign! m (mp/element-map m f a more))))
+    (element-reduce
+      ([m f]
+        (reduce f m))
+      ([m f init]
+        (reduce f init m))))
 
 (extend-protocol mp/PIndexedSetting
   DataSet
@@ -432,10 +566,14 @@
 ;; Printing methods for Datasets
 
 (defmethod print-dup DataSet [^clojure.core.matrix.impl.dataset.DataSet x ^Writer writer]
-  (.write writer (str "#dataset/dataset " (into {} x))))
+  (.write writer (str "#dataset/dataset " {:column-names (.column-names x)
+                                           :columns (.columns x)
+                                           :shape (.shape x)})))
 
 (defmethod print-method DataSet [^clojure.core.matrix.impl.dataset.DataSet x ^Writer writer]
-  (.write writer (str "#dataset/dataset " (into {} x))))
+  (.write writer (str "#dataset/dataset " {:column-names (.column-names x)
+                                           :columns (.columns x)
+                                           :shape (.shape x)})))
 
 (defmethod print-dup DataSetRow [^clojure.core.matrix.impl.dataset.DataSetRow x ^Writer writer]
   (.write writer (str "#dataset/row " {:column-names (.column-names x)